Agregar ruta para obtener transacciones recientes y refactorizar importaciones
